Fix DEBT-414 promo review findings (#729)
* Fix DEBT-414 promo review findings

* Strengthen legal link sanitization test
