+++ components/legal/legal-document.test.tsx
@@ -3,6 +3,7 @@ import { renderToStaticMarkup } from 'react-dom/server';
 import { beforeAll, describe, expect, it, vi } from 'vitest';
 import {
   findAnchorByHref,
+  findElementByText,
   findHeadingByText,
   parseHtml,
 } from '@/tests/shared/dom-helpers';
@@ -85,7 +86,9 @@ describe('LegalDocument', () => {
         }}
       />,
     );
+    const doc = parseHtml(html);
 
-    expect(html.toLowerCase()).not.toContain('javascript:');
+    const unsafeLink = findElementByText<HTMLAnchorElement>(doc, 'a', 'Unsafe');
+    expect(unsafeLink?.getAttribute('href') ?? '').not.toMatch(/^javascript:/i);
   });
 });
